FFmpeg preview encoding can now be done async Fix for delete_resource()-function since it didn't delete all files
git-svn-id: http://svn.resourcespace.org/svn/resourcespace@526 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/image_processing.php
@@ -752,8 +752,11 @@ function upload_preview($ref)
 	# Create previews
 	create_previews($ref,false,$extension,true);
 
-	# Delete temporary file.
-	unlink($filepath);
+	# Delete temporary file, if not transcoding.
+	if(!sql_value("SELECT is_transcoding value FROM resource WHERE ref = '".escape_check($ref)."'", false))
+		{
+		unlink($filepath);
+		}
 
     return true;
     }}
